name and comment parser done

+++ srcs/error.c
@@ -6,7 +6,7 @@
 /*   By: cdraugr- <cdraugr-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/09 13:42:23 by cdraugr-          #+#    #+#             */
-/*   Updated: 2019/10/20 22:00:16 by cdraugr-         ###   ########.fr       */
+/*   Updated: 2019/10/22 14:53:48 by cdraugr-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@ void	ft_error(char *message)
 	exit(1);
 }
 
-void	syntax_error(int32_t row, int32_t column)
+void	syntax_error(size_t row, size_t column)
 {
 	ft_putstr_fd("\033[1;31m", 2);
 	ft_putstr_fd("ERROR: Invalid syntax at line #", 2);
